requirements.txt

streamlit
pandas
google-generativeai
googletrans==4.0.0-rc1
python-dotenv
